+++ lifecompanion/lc-app/src/main/java/org/lifecompanion/model/impl/configurationcomponent/PredictionParameter.java
@@ -32,14 +32,14 @@ public class PredictionParameter implements PredictionParameterI {
     private StringProperty selectedCharPredictorId;
     private StringProperty charPredictionSpaceChar;
     private BooleanProperty enableMinWordPredictionScoreThreshold;
-    private FloatProperty minWordPredictionScoreThresholdProperty;
+    private DoubleProperty minWordPredictionScoreThresholdProperty;
 
     public PredictionParameter() {
         this.selectedCharPredictorId = new SimpleStringProperty(this, "selectedCharPredictorId", AutoCharPredictionController.INSTANCE.getDefaultPredictor().getId());
         this.selectedWordPredictorId = new SimpleStringProperty(this, "selectedWordPredictorId", WordPredictionController.INSTANCE.getDefaultPredictor().getId());
         this.charPredictionSpaceChar = new SimpleStringProperty(this, "charPredictionSpaceChar", "_");
         this.enableMinWordPredictionScoreThreshold = new SimpleBooleanProperty(this, "enableMinWordPredictionScoreThreshold",false);
-        this.minWordPredictionScoreThresholdProperty = new SimpleFloatProperty(this, "minWordPredictionScoreThreshold",(float)0.10);
+        this.minWordPredictionScoreThresholdProperty = new SimpleDoubleProperty(this, "minWordPredictionScoreThreshold",0.10);
     }
 
     @Override
@@ -61,7 +61,7 @@ public class PredictionParameter implements PredictionParameterI {
     public BooleanProperty enableMinWordPredictionScoreThresholdProperty() {
         return enableMinWordPredictionScoreThreshold;    }
 
-    public FloatProperty minWordPredictionScoreThresholdProperty(){
+    public DoubleProperty minWordPredictionScoreThresholdProperty(){
         return minWordPredictionScoreThresholdProperty;    }
 
     // Class part : "IO"
